+++ CHANGELOG.md
@@ -8,6 +8,13 @@
 
 ## pre-Beta 1.1
 - Added plugin init system
+- All plugin related commands are now in 'plugman' command
 - Started developing of unload function
 - Recompiled lib.zip
 - Added sample plugin
+
+## pre-Beta 1.2
+- Recompiled lip.zip
+- Unload function is now delayed
+- Added 'repoadd command'
+- Fixed typo in loader.py